fix: test case authenticate
