+++ middlewares/auth.js
@@ -1,0 +1,25 @@
+const passport = require('../config/passport')
+
+const authenticated = (req, res, next) => {
+  passport.authenticate('jwt', { session: false }, (err, user, info) => {
+    if (err) return next(err)
+    console.log(user)
+    if (!user) return next(new Error('no user'))
+    req.user = { ...user.dataValues }
+    return next()
+  })
+}
+
+const checkRole = (role = 'user') => {
+  return (req, res, next) => {
+    if (req.user.role !== role) {
+      return next(new Error('Permission denied.'))
+    }
+    return next()
+  }
+}
+
+module.exports = {
+  authenticated,
+  checkRole
+}
